pyproject.toml
[project]
name = "blackmusketeers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
    "streamlit>=1.56.0",
]
